+++ CHANGELOG.md
@@ -4,7 +4,7 @@ Yii2 Queue Extension Change Log
 2.3.7 under development
 -----------------------
 
-- no changes in this release.
+- Enh #412: Added path to an `.ini` file to a child process (werdender)
 
 
 2.3.6 October 03, 2023